feat: move FilterByListParams to commons for KMP
Move FilterByListParams and its core filtering logic to the commons
module to unblock KMP iOS migration for ~20+ DAL filters.

Changes:
- Create IFilterByListTopNavFilter interface in commons with minimal
  methods needed by FilterByListParams (matchAuthor, match, isGlobal,
  isMutedFilter, relayUrl)
- Move FilterByListParams class to commons/ui/dal/
- Make IFeedTopNavFilter extend IFilterByListTopNavFilter
- Add isGlobal override to GlobalTopNavFilter
- Add isMutedFilter override to MutedAuthors*TopNavFilter
- Add relayUrl override to RelayTopNavFilter
- Leave typealias at original location for backwards compatibility

All existing consumers continue to work unchanged via the typealias
and the IFeedTopNavFilter : IFilterByListTopNavFilter inheritance.
