+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/topNavFeeds/noteBased/muted/MutedAuthorsByOutboxTopNavFilter.kt
@@ -37,6 +37,8 @@ class MutedAuthorsByOutboxTopNavFilter(
     val authors: Set<String>,
     val blockedRelays: StateFlow<Set<NormalizedRelayUrl>>,
 ) : IFeedTopNavFilter {
+    override val isMutedFilter: Boolean get() = true
+
     override fun matchAuthor(pubkey: HexKey) = pubkey in authors
 
     override fun match(noteEvent: Event): Boolean =
