+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/topNavFeeds/noteBased/muted/MutedAuthorsByProxyTopNavFilter.kt
@@ -35,6 +35,8 @@ class MutedAuthorsByProxyTopNavFilter(
     val authors: Set<String>,
     val proxyRelays: Set<NormalizedRelayUrl>,
 ) : IFeedTopNavFilter {
+    override val isMutedFilter: Boolean get() = true
+
     override fun matchAuthor(pubkey: HexKey) = pubkey in authors
 
     override fun match(noteEvent: Event): Boolean =
